Refactored query features to separate files
